Migracion al base de datos postgresql aws. Cambio de tipo de datos para base de datos postresql

+++ accounts/src/main/java/com/eazybytes/accounts/mapper/AccountsMapper.java
@@ -6,14 +6,14 @@ import com.eazybytes.accounts.entity.Accounts;
 public class AccountsMapper {
 
     public static AccountsDto mapToAccountsDto(Accounts accounts, AccountsDto accountsDto) {
-        accountsDto.setAccountNumber(accounts.getAccountNumber());
+        accountsDto.setAccountNumber(Long.valueOf(accounts.getAccountNumber()));
         accountsDto.setAccountType(accounts.getAccountType());
         accountsDto.setBranchAddress(accounts.getBranchAddress());
         return accountsDto;
     }
 
     public static Accounts mapToAccounts(AccountsDto accountsDto, Accounts accounts) {
-        accounts.setAccountNumber(accountsDto.getAccountNumber());
+        accounts.setAccountNumber(Math.toIntExact(accountsDto.getAccountNumber()));
         accounts.setAccountType(accountsDto.getAccountType());
         accounts.setBranchAddress(accountsDto.getBranchAddress());
         return accounts;
